moved replays out of SQL tables and into storage, shorter emotes

+++ src/backend/services/mmr_service.py
@@ -24,8 +24,8 @@ class MMRService:
     """Helper for calculating new MMR values after a match."""
 
     _DEFAULT_MMR: int = 1500
-    _K_FACTOR: float = 40.0
-    _DIVISOR: float = 500.0
+    _K_FACTOR: int = 40
+    _DIVISOR: int = 500
 
     def calculate_new_mmr(self, player_one_mmr: int, player_two_mmr: int, result: int) -> MatchMMROutcome:
         """Return the updated MMR values for both players.
